refactor: move util/exec to own folder
Also moves related docker and env files
