+++ lib/manager/npm/post-update/index.ts
@@ -8,7 +8,7 @@ import * as lerna from './lerna';
 import * as yarn from './yarn';
 import * as pnpm from './pnpm';
 import * as hostRules from '../../../util/host-rules';
-import { getChildProcessEnv } from '../../../util/env';
+import { getChildProcessEnv } from '../../../util/exec/env';
 import { PostUpdateConfig, PackageFile, Upgrade } from '../../common';
 import { platform } from '../../../platform';
