+++ test/manager/mix/artifacts.spec.ts
@@ -4,7 +4,7 @@ import { platform as _platform } from '../../../lib/platform';
 import { updateArtifacts } from '../../../lib/manager/mix';
 import { mocked } from '../../util';
 import { envMock, mockExecAll } from '../../execUtil';
-import * as _env from '../../../lib/util/env';
+import * as _env from '../../../lib/util/exec/env';
 
 const fs: jest.Mocked<typeof _fs> = _fs as any;
 const exec: jest.Mock<typeof _exec> = _exec as any;
@@ -14,7 +14,7 @@ const env = mocked(_env);
 jest.mock('fs-extra');
 jest.mock('child_process');
 jest.mock('../../../lib/platform');
-jest.mock('../../../lib/util/env');
+jest.mock('../../../lib/util/exec/env');
 
 const config = {
   localDir: '/tmp/github/some/repo',
