+++ test/manager/pip_setup/extract.spec.ts
@@ -7,14 +7,14 @@ import {
   pythonVersions,
 } from '../../../lib/manager/pip_setup/extract';
 import { envMock, mockExecSequence } from '../../execUtil';
-import * as _env from '../../../lib/util/env';
+import * as _env from '../../../lib/util/exec/env';
 import { mocked } from '../../util';
 
 const exec: jest.Mock<typeof _exec> = _exec as any;
 const env = mocked(_env);
 
 jest.mock('child_process');
-jest.mock('../../../lib/util/env');
+jest.mock('../../../lib/util/exec/env');
 
 describe('lib/manager/pip_setup/extract', () => {
   beforeEach(() => {
