+++ test/manager/poetry/artifacts.spec.ts
@@ -4,11 +4,11 @@ import { updateArtifacts } from '../../../lib/manager/poetry/artifacts';
 import { platform as _platform } from '../../../lib/platform';
 import { mocked } from '../../util';
 import { envMock, mockExecAll } from '../../execUtil';
-import * as _env from '../../../lib/util/env';
+import * as _env from '../../../lib/util/exec/env';
 
 jest.mock('fs-extra');
 jest.mock('child_process');
-jest.mock('../../../lib/util/env');
+jest.mock('../../../lib/util/exec/env');
 
 const fs: jest.Mocked<typeof _fs> = _fs as any;
 const exec: jest.Mock<typeof _exec> = _exec as any;
