+++ test/util/env.spec.ts
@@ -1,4 +1,4 @@
-import { getChildProcessEnv } from '../../lib/util/env';
+import { getChildProcessEnv } from '../../lib/util/exec/env';
 
 describe('getChildProcess environment when trustlevel set to low', () => {
   const envVars = ['HTTP_PROXY', 'HTTPS_PROXY', 'NO_PROXY', 'HOME', 'PATH'];
